requirements.txt
eel==0.17.0
requests>=2.31.0
duckduckgo-search>=3.9.10,<4.0
chromadb>=0.5.0,<1.0
pymupdf4llm>=0.0.14
ollama>=0.3.0
beautifulsoup4>=4.12.0
aiohttp>=3.10.0
pyinstaller>=6.19.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
numpy>=1.24.0,<2.0
